myservice/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.8.1
app==0.0.1
APScheduler==3.6.3
asgiref==3.5.2
astroid==2.12.5
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
components==1.2.8
DateTime==4.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
Django==4.1.3
djangorestframework==3.14.0
entrypoints==0.4
ephem==4.1.3
exceptiongroup==1.0.0rc9
executing==1.0.0
flake8==6.0.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
frozenlist==1.3.3
greenlet==1.1.3
h11==0.14.0
httpie==3.2.1
idna==3.4
iniconfig==1.1.1
ipykernel==6.15.2
ipython==8.4.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
lazy-object-proxy==1.7.1
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.23.4
outcome==1.2.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.5
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.13.0
pylint==2.15.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.0.1
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pyzmq==23.2.1
repoze.lru==0.7
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
Routes==2.5.1
scipy==1.9.3
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.4.3
stack-data==0.5.0
termcolor==2.0.1
tomli==2.0.1
tomlkit==0.11.4
tornado==6.1
traitlets==5.3.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
yarl==1.8.2
zope.interface==5.4.0
